requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
alembic
pydantic
openai
google-generativeai==0.8.3
python-dotenv
httpx
aiosqlite
Pillow
pyjwt
python-multipart
cloudinary
